Exercises

+++ src/examples/Int.java
@@ -1,6 +1,6 @@
 package examples;
 
-public class Double {
+public class Int {
     public static void main(String[] args) {
         int a = 9 / 5;
         System.out.println(a);
